+++ src/Calculator/Operations/Basic/DivideOperation.cs
@@ -7,7 +7,7 @@ public class DivideOperation : BasicOperation
     {
     }
 
-    public override double Calculate()
+    public override decimal Calculate()
     {
         ArgumentNullException.ThrowIfNull(LeftOperand);
         ArgumentNullException.ThrowIfNull(RightOperand);
