+++ src/Calculator/Operations/Basic/PlusOperation.cs
@@ -7,7 +7,7 @@ public class PlusOperation : BasicOperation
     {
     }
 
-    public override double Calculate()
+    public override decimal Calculate()
     {
         ArgumentNullException.ThrowIfNull(LeftOperand);
         ArgumentNullException.ThrowIfNull(RightOperand);
